Use angular bindings for hovering and selecting
This commit also adds two new data structures for keeping track of the
state of the view.

One is tree that mirrors the structure of the main object, but also adds
state information used for highlighting and selecting. The zipper for
each part of the tree is also stored here.

The other is a flat list that points to the same objects as the
aforementioned tree. This is for easier lookup from the view.