Optimize view interactions with raw jQuery
View selection/hovering/navigation now uses a hand-bound event handler,
instead of adding ngClick/ngMouse*/ngClass directives to the HTML and
relying on the much slower $compile service.

The auto-scroll animation is no longer triggered by the ngClass
directive, so it is now done in the jsonData.select function.
Consequently, the ngAnimate module is not longer used, so it has been
removed.